Move Succeed State doctests to unit tests
